feat: remove unused doc
