+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './value-object'
+export * from './value-object';
