+++ src/value-object/index.ts
@@ -1,1 +1,1 @@
-export * from './primitives'
+export * from './primitives';
