+++ src/value-object/primitives/index.ts
@@ -1,4 +1,4 @@
-export * from './EnumValueObject'
-export * from './NumberValueObject'
-export * from './StringValueObject'
-export * from './DateValueObject'
+export * from './enum-value-object';
+export * from './number-value-object';
+export * from './string-value-object';
+export * from './date-value-object';
